Refactor Owner to FantasyTeams.Owner
* Refactor Owner to FantasyTeams.Owner
* See #797
