+++ lib/ex338_web/controllers/owner_controller.ex
@@ -1,7 +1,7 @@
 defmodule Ex338Web.OwnerController do
   use Ex338Web, :controller
 
-  alias Ex338.{FantasyLeagues.FantasyLeague, Owner}
+  alias Ex338.{FantasyLeagues.FantasyLeague, FantasyTeams.Owner}
 
   def index(conn, %{"fantasy_league_id" => league_id}) do
     fantasy_league = FantasyLeague |> Repo.get(league_id)
