+++ test/ex338/fantasy_teams_test.exs
@@ -1,6 +1,6 @@
 defmodule Ex338.FantasyTeamsTest do
   use Ex338.DataCase
-  alias Ex338.FantasyTeams
+  alias Ex338.{FantasyTeams, FantasyTeams.Owner}
 
   describe "count_pending_draft_queues/1" do
     test "returns number of pending draft queues for a team" do
@@ -202,6 +202,48 @@ defmodule Ex338.FantasyTeamsTest do
       [result] = FantasyTeams.find_owned_players(team.id)
 
       assert result.id == player_a.id
+    end
+  end
+
+  describe "get_email_recipients/2" do
+    test "return email addresses for a league" do
+      league_a = insert(:fantasy_league)
+      league_b = insert(:fantasy_league)
+      team_a = insert(:fantasy_team, team_name: "A", fantasy_league: league_a)
+      team_b = insert(:fantasy_team, team_name: "B", fantasy_league: league_b)
+      user_a = insert_user()
+      user_b = insert_user()
+      insert(:owner, fantasy_team: team_a, user: user_a)
+      insert(:owner, fantasy_team: team_b, user: user_b)
+
+      result = FantasyTeams.get_email_recipients_for_league(league_a.id)
+
+      assert result == [{user_a.name, user_a.email}]
+    end
+  end
+
+  describe "get_leagues_email_recipients/1" do
+    test "return email addresses for multiple leagues" do
+      league_a = insert(:fantasy_league)
+      league_b = insert(:fantasy_league)
+      team_a = insert(:fantasy_team, team_name: "A", fantasy_league: league_a)
+      team_b = insert(:fantasy_team, team_name: "B", fantasy_league: league_b)
+      user_a = insert_user()
+      user_b = insert_user()
+      _user_c = insert_user()
+      insert(:owner, fantasy_team: team_a, user: user_a)
+      insert(:owner, fantasy_team: team_b, user: user_b)
+
+      result =
+        FantasyTeams.get_leagues_email_addresses([
+          league_a.id,
+          league_b.id
+        ])
+
+      assert result == [
+               {user_b.name, user_b.email},
+               {user_a.name, user_a.email}
+             ]
     end
   end
 
@@ -517,6 +559,16 @@ defmodule Ex338.FantasyTeamsTest do
       assert result_a.team_name == team_a.team_name
       assert result_b.points == [0, 0, 0, 0, 0, 0, 8, 8, 8, 8, 8, 8]
       assert result_b.team_name == team_b.team_name
+    end
+  end
+
+  describe "update_owner/2" do
+    test "updates owner with valid attributes" do
+      owner = insert(:owner, rules: "unaccepted")
+      attrs = %{rules: "accepted"}
+
+      assert {:ok, %Owner{} = owner} = FantasyTeams.update_owner(owner, attrs)
+      assert owner.rules == :accepted
     end
   end
