+++ test/ex338_web/views/view_helpers_test.exs
@@ -8,7 +8,7 @@ defmodule Ex338Web.ViewHelpersViewTest do
     FantasyPlayers.FantasyPlayer,
     FantasyTeams.FantasyTeam,
     InSeasonDraftPick,
-    Owner,
+    FantasyTeams.Owner,
     RosterPositions.RosterPosition,
     Waivers.Waiver
   }
